refactor(cli): split trigger command helpers [OPE-361]
